requirements.txt
langgraph==0.4.3
langchain==0.3.25
langchain-anthropic==0.3.12
langchain-mcp-adapters==0.0.10
mcp==1.7.1
python-dotenv==1.1.0




